% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_SparseSet.R
\name{as_Set}
\alias{as_Set}
\title{Convert Named Vector to Set}
\usage{
as_Set(A)
}
\arguments{
\item{A}{A named vector or matrix to build a new \code{Set}.}
}
\value{
A \code{Set} object.
}
\description{
Convert Named Vector to Set
}
\examples{
A <- c(a = 0.1, b = 0.2, p = 0.3, q = 0)
as_Set(A)

}
