% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_SparseSet.R
\name{as_SparseSet}
\alias{as_SparseSet}
\title{Convert Named Vector to SparseSet}
\usage{
as_SparseSet(A)
}
\arguments{
\item{A}{A named vector or matrix to build a new \code{SparseSet}.}
}
\value{
A \code{SparseSet} object.
}
\description{
Convert Named Vector to SparseSet
}
\examples{
A <- c(a = 0.1, b = 0.2, p = 0.3, q = 0)
as_SparseSet(A)

}
