% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_density}
\alias{calculate_density}
\title{Calculate Fuzzy Density}
\usage{
calculate_density(extents, intents, I)
}
\arguments{
\item{extents}{A \code{SparseSet} or \code{CsparseMatrix} of extents.}

\item{intents}{A \code{SparseSet} or \code{CsparseMatrix} of intents.}

\item{I}{The original numeric matrix (FormalContext$I).}
}
\value{
A numeric vector.
}
\description{
Calculates the density of each concept in the original matrix I.
}
