% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_auxiliary.R
\name{calculate_grades}
\alias{calculate_grades}
\title{Calculate Concept Grades (Levels)}
\usage{
calculate_grades(concept_ids, edge_from, edge_to)
}
\arguments{
\item{concept_ids}{A vector of concept IDs (integers).}

\item{edge_from}{A vector of source concept IDs from the cover relation (Hasse diagram).}

\item{edge_to}{A vector of target concept IDs from the cover relation (Hasse diagram).}
}
\value{
An integer vector of the calculated grade for each concept ID.
}
\description{
Calculates the grade (level) of each concept using the longest path
from the bottom element. This is a fast C++ implementation.
}
