% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_miners_registry.R
\docType{data}
\name{conceptRegistry}
\alias{conceptRegistry}
\title{Concept Miners Registry}
\format{
An object of class \code{concept_miner_registry} (inherits from \code{registry}) of length 3.
}
\usage{
conceptRegistry
}
\description{
Concept Miners Registry
}
\details{
This is a registry that stores the concept miners that can be
applied using the \code{find_concepts()} method in an
\code{FormalConcept}.

One can obtain the list of available equivalence operators by:
\code{conceptRegistry$get_entry_names()}
}
\keyword{datasets}
