% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_from_file.R
\name{context_from_file}
\alias{context_from_file}
\title{Create a Context from File}
\usage{
context_from_file(filename)
}
\arguments{
\item{filename}{(character) The CSV o CXT file}
}
\value{
A context object
}
\description{
Create a Context from File
}
\examples{

filename <- system.file("digits.cxt", package = "fcaR")
cxt <- context_from_file(filename)

}
