% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_closure.R
\name{.compute_closure}
\alias{.compute_closure}
\title{Compute Closure of a (Fuzzy) Set}
\usage{
.compute_closure(S, LHS, RHS, attributes, reduce = FALSE,
  verbose = FALSE)
}
\arguments{
\item{S}{(a fuzzy set "dgCMatrix") Set to compute the closure.}

\item{LHS}{(dGCMatrix) Matrix of lhs of implications}

\item{attributes}{(character vector) Names of the attributes.}

\item{verbose}{}
}
\description{
This function is used to compute the closure...
}
