% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dgbasis_binary.R
\name{.get_dgbasis_binary_opt}
\alias{.get_dgbasis_binary_opt}
\title{Computes the Duquenne-Guigues basis using optimized Ganter's algorithm for crisp data.}
\usage{
.get_dgbasis_binary_opt(I, imp_basis = NULL, verbose = FALSE)
}
\description{
Computes the Duquenne-Guigues basis using optimized Ganter's algorithm for crisp data.
}
