% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_implications.R
\name{dplyr_implications}
\alias{dplyr_implications}
\alias{arrange.ImplicationSet}
\title{dplyr verbs for ImplicationSet}
\usage{
\method{arrange}{ImplicationSet}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{ImplicationSet}.}

\item{...}{Expressions used for filtering, arranging, or slicing.}
}
\description{
S3 methods to enable dplyr verbs on ImplicationSet objects.
}
