% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_implications.R
\name{dplyr_rules}
\alias{dplyr_rules}
\alias{arrange.RuleSet}
\title{dplyr verbs for RuleSet}
\usage{
\method{arrange}{RuleSet}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{RuleSet} or \code{ImplicationSet}.}

\item{...}{Expressions used for filtering, arranging, or slicing.}
}
\description{
S3 methods to enable dplyr verbs on RuleSet (and ImplicationSet) objects.
}
