% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{dplyr_verbs}
\alias{dplyr_verbs}
\alias{select.FormalContext}
\title{dplyr verbs for FormalContext}
\usage{
\method{select}{FormalContext}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{FormalContext}.}

\item{...}{Arguments passed to the corresponding dplyr verbs.}
}
\description{
S3 methods to enable dplyr verbs on FormalContext objects.
}
