% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_concepts.R
\name{equal_implications}
\alias{equal_implications}
\title{Equality of sets of implications}
\usage{
equal_implications(imps1, imps2)
}
\arguments{
\item{imps1}{(\code{RuleSet} or \code{ImplicationSet}) The first set of implications.}

\item{imps2}{(\code{RuleSet} or \code{ImplicationSet}) The second set of implications.}
}
\value{
\code{TRUE} if both sets contain the exact same implications, \code{FALSE} otherwise.
}
\description{
Tests whether two \code{ImplicationSet} or \code{RuleSet} objects represent
the exact same set of implications, regardless of the order in which they appear.
It uses a prefix tree (Trie) over the premises (LHS) for fast matching.
}
\examples{
fc <- FormalContext$new(planets)
fc$find_implications()
imps1 <- fc$implications$clone()
# Shuffle the implications
imps2 <- imps1[sample(imps1$cardinality())]
equal_implications(imps1, imps2)
imps1 \%==\% imps2
}
