% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalences_registry.R
\docType{data}
\name{equivalencesRegistry}
\alias{equivalencesRegistry}
\title{Equivalence Rules Registry}
\format{
An object of class \code{equivalence_registry} (inherits from \code{registry}) of length 6.
}
\usage{
equivalencesRegistry
}
\description{
Equivalence Rules Registry
}
\details{
This is a registry that stores the equivalence rules that can be
applied using the \code{apply_rules()} method in an
\code{ImplicationSet}.

One can obtain the list of available equivalence operators by:
\code{equivalencesRegistry$get_entry_names()}
}
\keyword{datasets}
