% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{fcaR_options}
\alias{fcaR_options}
\title{Set or get options for fcaR}
\usage{
fcaR_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for fcaR
}
\section{Supported options}{

The following options are supported
\itemize{
\item{\code{decimal_places}}{(\code{numeric};2) The number of decimal places to show when printing or exporting to \LaTeX sets, implications, concepts, etc.}
\item{\code{latex_size}}{(\code{character};"normalsize") Size to use when exporting to LaTeX.}
}
}

