% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.R
\name{find_causal_rules}
\alias{find_causal_rules}
\title{Causal Association Rules}
\usage{
find_causal_rules(
  fc,
  response_var,
  min_support = 0.1,
  confidence_level = 0.95,
  max_length = 3,
  verbose = FALSE
)
}
\arguments{
\item{fc}{A FormalContext object.}

\item{response_var}{(character) The name of the response variable.}

\item{min_support}{(numeric) Minimum support.}

\item{confidence_level}{(numeric) Confidence level for the causality test.}

\item{max_length}{(integer) Maximum length of the premise.}

\item{verbose}{(logical) Show verbose output?}
}
\value{
A data frame with causal rules and their quality metrics.
}
\description{
Functions to mine causal association rules from a formal context.
}
