% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formal_context.R
\docType{data}
\name{formal_context}
\alias{formal_context}
\title{R6 class for a formal context}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
formal_context
}
\arguments{
\item{I}{(numeric matrix) The table of the formal context.}

\item{grades_set}{(numeric vector, optional) the anumeration of the grades of the attributes.}

\item{remove_const}{(logical) If \code{TRUE}, remove constant columns. The default is \code{FALSE}.}

\item{impl_set}{(\code{ImplicationSet} object) The implications to add to this formal context.}

\item{verbose}{(logical) TRUE will provide a verbose output.}

\item{verbose}{(logical) TRUE will provide a verbose output.}

\item{quality}{(logical) Compute the interest measures for each rule?}

\item{minsupp}{(numeric) Only plot the sublattice generated with those concepts whose support is greater than or equal to \code{minsupp}.}

\item{object_names}{(logical) If \code{TRUE}, plot object names, otherwise omit them from the diagram.}
}
\value{
An object of the \code{FormalContext} class.

Nothing, just updates the internal \code{implications} field.

A list with all the concepts in the formal context.

Nothing, just updates the internal fields \code{concepts} and \code{implications}.

A \code{transactions} object.

A \code{rules} object from \code{arules} package. If \code{quality == TRUE}, the function \code{interestMeasure} from \code{arules} is used.

A vector with (number of objects, number of attributes).

Prints information regarding the formal context.

Nothing, just plots the graph of the concept lattice.

Nothing, just plots the formal context.

A vector with the support of each concept.

A vector with the support of each implication
}
\description{
This class implements the data structure and methods for formal contexts.

Creator for the Fomal Context class

Add a precomputed implication set

Use Ganter Algorithm to compute concepts

Use modified Ganter algorithm to compute both concepts and implications

Convert the formal context to object of class \code{transactions} from the \code{arules} package

Export implications to \code{arules} format

Dimensions of the formal context

Prints the formal context

Plot the concept lattice

Plot the formal context table

Get support of each concept

Compute support of each implication
}
\details{
Columns of \code{I} must be named, and are the names of the attributes of the formal context.
}
\section{Public fields}{

\itemize{
\item \code{I}: the table for the formal context.
\item \code{attributes}: name of the attributes in the formal context.
\item \code{objects}: name of the objects in the context.
\item \code{grades_set}: set of grades (in \link{0, 1}) of the attributes.
\item \code{concepts}: list of concepts (extent, intent).
\item \code{implications}: extracted implications as an \code{ImplicationSet}.
\item \code{concept_support}: vector with the support of the concepts.
\item \code{implications_support}: vector with support of the extracted implications.
}
}

\keyword{datasets}
