% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intension_extension.R
\name{get_extension}
\alias{get_extension}
\title{Extension of Properties}
\usage{
get_extension(cxt, obj_list)
}
\arguments{
\item{cxt}{Contexts where to apply intension}

\item{prop_list}{List or vector with properties names}
}
\value{
The extension (derivation) of the given properties
}
\description{
Extension of Properties
}
\examples{
cxt_file <- system.file("digits.cxt", package = "fcaR")
cxt <- context_from_file(cxt_file)
cxt \%>\% get_extension(c("c", "d"))

}
