% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{get_fcarepository_contexts}
\alias{get_fcarepository_contexts}
\title{Get Metadata from the FCA Repository}
\usage{
get_fcarepository_contexts()
}
\value{
A list containing the metadata for each context (title, dimensions, description, source).
}
\description{
Connects to the official FCA Repository (https://fcarepository.org) and downloads
the metadata for all available datasets.
}
\examples{
\dontrun{
meta <- get_fcarepository_contexts()
}
}
