% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intension_extension.R
\name{get_intension}
\alias{get_intension}
\title{Intension of Objects}
\usage{
get_intension(cxt, obj_list)
}
\arguments{
\item{cxt}{Contexts where to apply intension}

\item{obj_list}{List or vector with object names}
}
\value{
The intension (derivation) of the given objects
}
\description{
Intension of Objects
}
\examples{
cxt_file <- system.file("digits.cxt", package = "fcaR")
cxt <- context_from_file(cxt_file)
cxt \%>\% get_intension(c("0", "8"))

}
