% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holds_respects.R
\name{\%holds_in\%}
\alias{\%holds_in\%}
\title{Implications that hold in a Formal Context}
\usage{
imps \%holds_in\% fc
}
\arguments{
\item{imps}{(\code{ImplicationSet}) The set of implications to test if hold in the formal context.}

\item{fc}{(\code{FormalContext}) A formal context where to test if the implications hold.}
}
\value{
A logical vector, indicating if each implication holds in the formal context.
}
\description{
Implications that hold in a Formal Context
}
\examples{
fc <- FormalContext$new(planets)
fc$find_implications()
imps <- fc$implications$clone()
imps \%holds_in\% fc

}
