% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entails_equiv.R
\name{\%~\%}
\alias{\%~\%}
\title{Equivalence of sets of implications}
\usage{
imps \%~\% imps2
}
\arguments{
\item{imps}{A \code{ImplicationSet}.}

\item{imps2}{Another \code{ImplicationSet}.}
}
\value{
\code{TRUE} of and only if \code{imps} and \code{imps2}
are equivalent, that is, if every implication in \code{imps}
follows from \code{imps2} and viceversa.
}
\description{
Equivalence of sets of implications
}
\examples{
fc <- FormalContext$new(planets)
fc$find_implications()
imps <- fc$implications$clone()
imps2 <- imps$clone()
imps2$apply_rules(c("simp", "rsimp"))
imps \%~\% imps2
imps \%~\% imps2[1:9]
}
