% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{guesswho}
\alias{guesswho}
\title{Data for Guess Who Board Game}
\format{
A matrix with 24 rows and 19 columns. Column names correspond to the physical attributes:
\describe{
\item{Gender}{Male, Female.}
\item{Hair Color}{BlackHair, WhiteHair, BlondeHair, BrownHair, RedHair.}
\item{Baldness}{Bald.}
\item{Eye Color}{BlueEyes, BrownEyes.}
\item{Facial Hair}{Moustache, Beard.}
\item{Accessories}{Hat, Glasses, Earrings.}
\item{Other features}{RosyCheeks, BigNose, BigMouth, Sad.}
}

Each entry contains a 1 if the character possesses the trait and a 0 otherwise.
}
\source{
Extracted from the physical version of the Guess Who board game characters.
}
\usage{
guesswho
}
\description{
A dataset representing the 24 classical characters from the Guess Who board game with their 19 binary attributes.
}
\keyword{datasets}
