% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implication_set.R
\name{implications_from_json}
\alias{implications_from_json}
\title{Import ImplicationSet from JSON}
\usage{
implications_from_json(json_str)
}
\arguments{
\item{json_str}{A JSON string generated by \code{to_json()}.}
}
\value{
An \code{ImplicationSet} object.
}
\description{
Reconstructs an ImplicationSet object from a JSON string.
}
