% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{install_concepts}
\alias{install_concepts}
\title{Install Python `'Concepts' Framework}
\usage{
install_concepts()
}
\value{
The 'concepts' module
}
\description{
Install Python `'Concepts' Framework
}
