% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonds.R
\name{is_bond}
\alias{is_bond}
\title{Verify if a relation is a bond between two formal contexts}
\usage{
is_bond(fc1, fc2, relation)
}
\arguments{
\item{fc1}{(\code{FormalContext}) The first formal context.}

\item{fc2}{(\code{FormalContext}) The second formal context.}

\item{relation}{(matrix) A binary matrix or \code{FormalContext} representing the relation between objects of \eqn{fc1} and attributes of \eqn{fc2}.}
}
\value{
\code{TRUE} if the relation is a bond, \code{FALSE} otherwise.
}
\description{
A bond between two formal contexts \eqn{K_1 = (G_1, M_1, I_1)} and \eqn{K_2 = (G_2, M_2, I_2)}
is a relation \eqn{R \subseteq G_1 \times M_2} such that every row of \eqn{R} is an intent
of \eqn{K_2} and every column of \eqn{R} is an extent of \eqn{K_1}.
}
