% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_plot.R
\name{lattice_plot}
\alias{lattice_plot}
\title{Plot Concept Lattice}
\usage{
lattice_plot(
  nodes_df,
  cover_matrix,
  method = "sugiyama",
  mode = NULL,
  objects = NULL,
  attributes = NULL,
  object_names = TRUE,
  to_latex = FALSE,
  extents = NULL,
  intents = NULL,
  ...
)
}
\arguments{
\item{nodes_df}{Data frame with 'id'.}

\item{cover_matrix}{Sparse matrix.}

\item{method}{Layout method ("sugiyama", "force").}

\item{mode}{Labeling mode ("reduced", "full", "empty").}

\item{objects}{Character vector.}

\item{attributes}{Character vector.}

\item{object_names}{Logical (Deprecated).}

\item{to_latex}{Logical.}

\item{extents}{List of extents.}

\item{intents}{List of intents.}

\item{...}{Extra args.}
}
\description{
Visualization of the concept lattice using 'ggraph'.
}
