% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcaR_helpers.R
\name{object_set}
\alias{object_set}
\title{Build an object \code{Set} from names}
\usage{
object_set(fc, ...)
}
\arguments{
\item{fc}{A \code{FormalContext}.}

\item{...}{Object names: a character vector, or several strings.}
}
\value{
A \code{Set} over \code{fc$objects}.
}
\description{
Dual of \code{attribute_set()}: builds a \code{Set} over \code{fc$objects}.
}
\examples{
\dontrun{
object_set(fc, "Squid Game", "Dark")
}

}
