% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_implication}
\alias{parse_implication}
\title{Parses a string into an implication}
\usage{
parse_implication(string, attributes)
}
\arguments{
\item{string}{(character) The string to be parsed}

\item{attributes}{(character vector) The attributes' names}
}
\value{
Two vectors as sparse matrices representing the LHS and RHS of the implication
}
\description{
Parses a string into an implication
}
