% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptlist.R
\name{print.conceptlist}
\alias{print.conceptlist}
\title{Print a list of concepts}
\usage{
\method{print}{conceptlist}(x, ...)
}
\arguments{
\item{x}{A list of concepts as a result of subsetting a \code{ConceptLattice}}
}
\value{
Nothing, just prints a list of concepts nicely.
}
\description{
Print a list of concepts
}
