% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_exporter.R
\name{print.tikz_code}
\alias{print.tikz_code}
\title{Print TikZ Code}
\usage{
\method{print}{tikz_code}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tikz_code}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Prints the TikZ code to the console.
}
