% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpM3.R
\name{remove_columns}
\alias{remove_columns}
\title{#' @export
#' @noRd
\verb{[.SpM} <- function(x, ...) x \%>\% extract_columns(...)}
\usage{
remove_columns(private, idx)
}
\description{
#' @export
#' @noRd
\verb{[.SpM} <- function(x, ...) x \%>\% extract_columns(...)
}
