% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_set.R
\name{rules_from_json}
\alias{rules_from_json}
\title{Import RuleSet from JSON}
\usage{
rules_from_json(json_str)
}
\arguments{
\item{json_str}{A JSON string generated by \code{to_json()}.}
}
\value{
A \code{RuleSet} object.
}
\description{
Reconstructs a RuleSet object from a JSON string.
}
