% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcaRviz.R
\name{run_fcaRviz}
\alias{run_fcaRviz}
\title{Launch the fcaRviz Interactive Viewer}
\usage{
run_fcaRviz(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny]{runApp}}.}
}
\description{
This function opens the interactive Shiny application fcaRviz for visual exploration.
}
\examples{
\dontrun{
run_fcaRviz()
}
}
