% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_exporter.R
\name{save_tikz}
\alias{save_tikz}
\title{Save TikZ Code to File}
\usage{
save_tikz(x, file, ...)
}
\arguments{
\item{x}{An object of class 'tikz_code'.}

\item{file}{Character. Path to the output file.}

\item{...}{Additional arguments.}
}
\description{
Exports the generated TikZ code to a .tex file.
}
