% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_registry.R
\docType{data}
\name{scalingRegistry}
\alias{scalingRegistry}
\title{Scaling Registry}
\format{
An object of class \code{scaling_registry} (inherits from \code{registry}) of length 5.
}
\usage{
scalingRegistry
}
\description{
Scaling Registry
}
\details{
This is a registry that stores the implemented scales that can be
applied using the \code{scale()} method in an
\code{FormalContext}.

One can obtain the list of available equivalence operators by:
\code{scalingRegistry$get_entry_names()}
}
\keyword{datasets}
