% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_concept.R
\docType{data}
\name{sparse_concept}
\alias{sparse_concept}
\title{R6 class for a fuzzy concept with sparse internal representation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sparse_concept
}
\arguments{
\item{extent}{(\code{SparseSet}) The extent of the concept.}

\item{intent}{(\code{SparseSet}) The intent of the concept.}
}
\value{
An object of class \code{SparseConcept}.

The \code{SparseSet} representation of the extent.

The \code{SparseSet} representation of the intent.

A string with the elements of the set and their grades between brackets {}.

The fuzzy concept in LaTeX.
}
\description{
This class implements the data structure and methods for fuzzy concepts.

Creator for objects of class \code{SparseConcept}

Internal \code{SparseSet} for the extent

Internal \code{SparseSet} for the intent

Prints the concept to console

Write the concept in LaTeX format
}
\keyword{datasets}
