% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_set.R
\name{sparse_set}
\alias{sparse_set}
\title{R6 class for a fuzzy set with sparse internal representation}
\description{
This class implements the data structure and methods for fuzzy sets.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{sparse_set$new()}}
\item \href{#method-assign}{\code{sparse_set$assign()}}
\item \href{#method-get_vector}{\code{sparse_set$get_vector()}}
\item \href{#method-get_attributes}{\code{sparse_set$get_attributes()}}
\item \href{#method-length}{\code{sparse_set$length()}}
\item \href{#method-print}{\code{sparse_set$print()}}
\item \href{#method-to_latex}{\code{sparse_set$to_latex()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_set$new(attributes, M = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign"></a>}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_set$assign(attributes, values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_vector"></a>}}
\subsection{Method \code{get_vector()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_set$get_vector()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_attributes"></a>}}
\subsection{Method \code{get_attributes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_set$get_attributes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-length"></a>}}
\subsection{Method \code{length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_set$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_set$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_latex"></a>}}
\subsection{Method \code{to_latex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_set$to_latex()}\if{html}{\out{</div>}}
}

}
}
