% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_cxt.R
\name{to_cxt}
\alias{to_cxt}
\title{Export to CXT format}
\usage{
to_cxt(I, objects, attributes, filename)
}
\arguments{
\item{I}{(Binary) Incidence matrix}

\item{objects}{Objects of the context}

\item{attributes}{Attributes of the context}

\item{filename}{Path to the file where to save the CXT format.}
}
\value{
\code{TRUE} invisibly.
}
\description{
Export to CXT format
}
