% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptlist.R
\name{to_latex}
\alias{to_latex}
\title{Print a list of concepts to LaTeX}
\usage{
to_latex(x)
}
\arguments{
\item{x}{A \code{conceptlist} as result of subsetting a \code{ConceptLattice}}
}
\value{
Nothing, just prints a list of concepts in LaTeX nicely.
}
\description{
Print a list of concepts to LaTeX
}
