% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_lattice.R
\name{BondLattice}
\alias{BondLattice}
\title{R6 class for a Bond Lattice}
\description{
This class extends \code{ConceptLattice} to represent the lattice of bonds between
two formal contexts \eqn{K_1} and \eqn{K_2}.
It inherits all lattice operations (finding irreducible elements, subconcepts, etc.)
and provides specific methods to extract bonds as matrices and FormalContexts.
}
\examples{
\donttest{
set.seed(42)
mat1 <- matrix(sample(0:1, 15, replace = TRUE), nrow = 5, ncol = 3)
rownames(mat1) <- paste0("O", 1:5)
colnames(mat1) <- paste0("A", 1:3)
fc1 <- FormalContext$new(mat1)

mat2 <- matrix(sample(0:1, 12, replace = TRUE), nrow = 4, ncol = 3)
rownames(mat2) <- paste0("P", 1:4)
colnames(mat2) <- paste0("B", 1:3)
fc2 <- FormalContext$new(mat2)

bl <- bonds(fc1, fc2)

# Extract all bonds as FormalContext objects
my_bonds <- bl$get_bonds()
print(my_bonds[[1]])
}

}
\section{Super classes}{
\code{\link[fcaR:ConceptSet]{fcaR::ConceptSet}} -> \code{\link[fcaR:ConceptLattice]{fcaR::ConceptLattice}} -> \code{BondLattice}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{elapsed}}{(numeric) Time elapsed during bond computation (C++ level)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BondLattice-new}{\code{BondLattice$new()}}
\item \href{#method-BondLattice-get_bonds}{\code{BondLattice$get_bonds()}}
\item \href{#method-BondLattice-similarity}{\code{BondLattice$similarity()}}
\item \href{#method-BondLattice-get_core}{\code{BondLattice$get_core()}}
\item \href{#method-BondLattice-print}{\code{BondLattice$print()}}
\item \href{#method-BondLattice-is_bond}{\code{BondLattice$is_bond()}}
\item \href{#method-BondLattice-clone}{\code{BondLattice$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptSet" data-id="["><a href='../../fcaR/html/ConceptSet.html#method-ConceptSet-bracket'><code>fcaR::ConceptSet$[()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptSet" data-id="extents"><a href='../../fcaR/html/ConceptSet.html#method-ConceptSet-extents'><code>fcaR::ConceptSet$extents()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptSet" data-id="intents"><a href='../../fcaR/html/ConceptSet.html#method-ConceptSet-intents'><code>fcaR::ConceptSet$intents()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptSet" data-id="is_empty"><a href='../../fcaR/html/ConceptSet.html#method-ConceptSet-is_empty'><code>fcaR::ConceptSet$is_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptSet" data-id="size"><a href='../../fcaR/html/ConceptSet.html#method-ConceptSet-size'><code>fcaR::ConceptSet$size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptSet" data-id="sub"><a href='../../fcaR/html/ConceptSet.html#method-ConceptSet-sub'><code>fcaR::ConceptSet$sub()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptSet" data-id="support"><a href='../../fcaR/html/ConceptSet.html#method-ConceptSet-support'><code>fcaR::ConceptSet$support()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptSet" data-id="to_latex"><a href='../../fcaR/html/ConceptSet.html#method-ConceptSet-to_latex'><code>fcaR::ConceptSet$to_latex()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptSet" data-id="to_list"><a href='../../fcaR/html/ConceptSet.html#method-ConceptSet-to_list'><code>fcaR::ConceptSet$to_list()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="bottom"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-bottom'><code>fcaR::ConceptLattice$bottom()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="decompose"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-decompose'><code>fcaR::ConceptLattice$decompose()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="density"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-density'><code>fcaR::ConceptLattice$density()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="dimension"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-dimension'><code>fcaR::ConceptLattice$dimension()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="infimum"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-infimum'><code>fcaR::ConceptLattice$infimum()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="is_atomic"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-is_atomic'><code>fcaR::ConceptLattice$is_atomic()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="is_distributive"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-is_distributive'><code>fcaR::ConceptLattice$is_distributive()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="is_modular"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-is_modular'><code>fcaR::ConceptLattice$is_modular()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="is_semimodular"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-is_semimodular'><code>fcaR::ConceptLattice$is_semimodular()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="join_irreducibles"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-join_irreducibles'><code>fcaR::ConceptLattice$join_irreducibles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="lower_neighbours"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-lower_neighbours'><code>fcaR::ConceptLattice$lower_neighbours()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="meet_irreducibles"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-meet_irreducibles'><code>fcaR::ConceptLattice$meet_irreducibles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="plot"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-plot'><code>fcaR::ConceptLattice$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="separation"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-separation'><code>fcaR::ConceptLattice$separation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="set_state"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-set_state'><code>fcaR::ConceptLattice$set_state()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="stability"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-stability'><code>fcaR::ConceptLattice$stability()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="subconcepts"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-subconcepts'><code>fcaR::ConceptLattice$subconcepts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="sublattice"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-sublattice'><code>fcaR::ConceptLattice$sublattice()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="sublattice_from"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-sublattice_from'><code>fcaR::ConceptLattice$sublattice_from()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="superconcepts"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-superconcepts'><code>fcaR::ConceptLattice$superconcepts()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="supremum"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-supremum'><code>fcaR::ConceptLattice$supremum()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="to_json"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-to_json'><code>fcaR::ConceptLattice$to_json()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="top"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-top'><code>fcaR::ConceptLattice$top()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="upper_neighbours"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-upper_neighbours'><code>fcaR::ConceptLattice$upper_neighbours()</code></a></span></li>
<li><span class="pkg-link" data-pkg="fcaR" data-topic="ConceptLattice" data-id="width"><a href='../../fcaR/html/ConceptLattice.html#method-ConceptLattice-width'><code>fcaR::ConceptLattice$width()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BondLattice-new"></a>}}
\if{latex}{\out{\hypertarget{method-BondLattice-new}{}}}
\subsection{Method \code{new()}}{
Initialize a BondLattice object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BondLattice$new(extents, intents, objects, attributes, I, fc1, fc2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{extents}}{(dgCMatrix) The extents of all concepts}

\item{\code{intents}}{(dgCMatrix) The intents of all concepts}

\item{\code{objects}}{(character) Names of the objects (not required for bonds, can be generic)}

\item{\code{attributes}}{(character) Names of the attributes (flattened $G_1 \\times M_2$)}

\item{\code{I}}{(matrix) Built incidence matrix}

\item{\code{fc1}}{(FormalContext) The first formal context}

\item{\code{fc2}}{(FormalContext) The second formal context}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BondLattice} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BondLattice-get_bonds"></a>}}
\if{latex}{\out{\hypertarget{method-BondLattice-get_bonds}{}}}
\subsection{Method \code{get_bonds()}}{
Extract the bonds represented by the intents of the lattice.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BondLattice$get_bonds(indices = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indices}}{(numeric or logical vector) The indices of the bonds (concepts) to extract. If \code{NULL} (default), extracts all.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of \code{FormalContext} objects, each representing one bond.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BondLattice-similarity"></a>}}
\if{latex}{\out{\hypertarget{method-BondLattice-similarity}{}}}
\subsection{Method \code{similarity()}}{
Compute similarity, affinity, or complexity metrics between the two contexts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BondLattice$similarity(
  type = c("log-bond", "top-density", "complexity", "core-agreement", "entropy",
    "stability", "width", "dimension", "width-index", "dimension-index")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{(character) The type of metric to compute:
\itemize{
\item \code{"log-bond"}: (Default) Normalized log-ratio of bonds. High value means high logical affinity.
\item \code{"top-density"}: Density of the largest possible bond (the top of the lattice).
\item \code{"complexity"}: Ratio of irreducible bonds to total bonds. Low value implies high structural emergence.
\item \code{"core-agreement"}: Ratio of filled cells in the Core bond versus the Top bond. Measures fundamental consensus.
\item \code{"entropy"}: Interaction entropy based on the log-size of the lattices.
\item \code{"stability"}: Average stability of the bonds in the lattice.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value representing the metric.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BondLattice-get_core"></a>}}
\if{latex}{\out{\hypertarget{method-BondLattice-get_core}{}}}
\subsection{Method \code{get_core()}}{
Get the 'Core' bond (the smallest possible bond).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BondLattice$get_core()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{FormalContext} representing the minimal bond.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BondLattice-print"></a>}}
\if{latex}{\out{\hypertarget{method-BondLattice-print}{}}}
\subsection{Method \code{print()}}{
Print the BondLattice object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BondLattice$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, just prints the object summary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BondLattice-is_bond"></a>}}
\if{latex}{\out{\hypertarget{method-BondLattice-is_bond}{}}}
\subsection{Method \code{is_bond()}}{
Verify if a relation is a bond between the internal contexts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BondLattice$is_bond(relation)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{relation}}{(matrix or FormalContext) The relation to verify.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if it's a bond, FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BondLattice-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BondLattice-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BondLattice$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
