% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formal_context.R
\name{formal_context}
\alias{formal_context}
\title{R6 class for a formal context}
\description{
This class implements the data structure and methods for formal contexts.
}
\section{Public fields}{

\itemize{
\item \code{I}: the table for the formal context.
\item \code{attributes}: name of the attributes in the formal context.
\item \code{objects}: name of the objects in the context.
\item \code{grades_set}: set of grades (in \link{0, 1}) of the attributes.
\item \code{concepts}: list of concepts (extent, intent).
\item \code{implications}: extracted implications as an \code{ImplicationSet}.
\item \code{concept_support}: vector with the support of the concepts.
\item \code{implications_support}: vector with support of the extracted implications.
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{formal_context$new()}}
\item \href{#method-add_implications}{\code{formal_context$add_implications()}}
\item \href{#method-compute_concepts}{\code{formal_context$compute_concepts()}}
\item \href{#method-extract_implications_concepts}{\code{formal_context$extract_implications_concepts()}}
\item \href{#method-convert_to_transactions}{\code{formal_context$convert_to_transactions()}}
\item \href{#method-export_implications_to_arules}{\code{formal_context$export_implications_to_arules()}}
\item \href{#method-dim}{\code{formal_context$dim()}}
\item \href{#method-print}{\code{formal_context$print()}}
\item \href{#method-plot_lattice}{\code{formal_context$plot_lattice()}}
\item \href{#method-plot_context}{\code{formal_context$plot_context()}}
\item \href{#method-get_concept_support}{\code{formal_context$get_concept_support()}}
\item \href{#method-get_implication_support}{\code{formal_context$get_implication_support()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creator for the Fomal Context class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$new(
  I,
  grades_set = sort(unique(as.vector(I))),
  remove_const = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{I}}{(numeric matrix) The table of the formal context.}

\item{\code{grades_set}}{(numeric vector, optional) the anumeration of the grades of the attributes.}

\item{\code{remove_const}}{(logical) If \code{TRUE}, remove constant columns. The default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Columns of \code{I} must be named, and are the names of the attributes of the formal context.
}

\subsection{Returns}{
An object of the \code{FormalContext} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_implications"></a>}}
\subsection{Method \code{add_implications()}}{
Add a precomputed implication set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$add_implications(impl_set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{impl_set}}{(\code{ImplicationSet} object) The implications to add to this formal context.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, just updates the internal \code{implications} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_concepts"></a>}}
\subsection{Method \code{compute_concepts()}}{
Use Ganter Algorithm to compute concepts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$compute_concepts(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{(logical) TRUE will provide a verbose output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with all the concepts in the formal context.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_implications_concepts"></a>}}
\subsection{Method \code{extract_implications_concepts()}}{
Use modified Ganter algorithm to compute both concepts and implications
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$extract_implications_concepts(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{(logical) TRUE will provide a verbose output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, just updates the internal fields \code{concepts} and \code{implications}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert_to_transactions"></a>}}
\subsection{Method \code{convert_to_transactions()}}{
Convert the formal context to object of class \code{transactions} from the \code{arules} package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$convert_to_transactions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{transactions} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_implications_to_arules"></a>}}
\subsection{Method \code{export_implications_to_arules()}}{
Export implications to \code{arules} format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$export_implications_to_arules(quality = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quality}}{(logical) Compute the interest measures for each rule?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{rules} object from \code{arules} package. If \code{quality == TRUE}, the function \code{interestMeasure} from \code{arules} is used.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dim"></a>}}
\subsection{Method \code{dim()}}{
Dimensions of the formal context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$dim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector with (number of objects, number of attributes).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Prints the formal context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Prints information regarding the formal context.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_lattice"></a>}}
\subsection{Method \code{plot_lattice()}}{
Plot the concept lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$plot_lattice()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, just plots the graph of the concept lattice.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_context"></a>}}
\subsection{Method \code{plot_context()}}{
Plot the formal context table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$plot_context()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, just plots the formal context.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_concept_support"></a>}}
\subsection{Method \code{get_concept_support()}}{
Get support of each concept
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$get_concept_support()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector with the support of each concept.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_implication_support"></a>}}
\subsection{Method \code{get_implication_support()}}{
Compute support of each implication
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$get_implication_support()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector with the support of each implication
}
}
}
