% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formal_context.R
\name{FormalContext}
\alias{FormalContext}
\title{R6 class for a formal context}
\description{
This class implements the data structure and methods for formal contexts.
}
\section{Public fields}{

\itemize{
\item \code{I}: the table for the formal context.
\item \code{attributes}: name of the attributes in the formal context.
\item \code{objects}: name of the objects in the context.
\item \code{grades_set}: set of grades (in [0, 1]) of the attributes.
\item \code{concepts}: list of concepts (extent, intent).
\item \code{implications}: extracted implications as an \code{ImplicationSet}.
\item \code{concept_support}: vector with the support of the concepts.
\item \code{implications_support}: vector with support of the extracted implications.
}
}

\references{
Guigues J, Duquenne V (1986). “Familles minimales d'implications informatives résultant d'un tableau de données binaires.” \emph{Mathématiques et Sciences humaines}, \emph{95}, 5-18.

Ganter B, Wille R (1999). \emph{Formal concept analysis : mathematical foundations}. Springer. ISBN 3540627715.

Belohlavek R (2002). “Algorithms for fuzzy concept lattices.” In \emph{Proc. Fourth Int. Conf. on Recent Advances in Soft Computing}. Nottingham, United Kingdom, 200-205.

Hahsler M, Grun B, Hornik K (2005). “arules - a computational environment for mining association rules and frequent item sets.” \emph{J Stat Softw}, \emph{14}, 1-25.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{formal_context$new()}}
\item \href{#method-is_empty}{\code{formal_context$is_empty()}}
\item \href{#method-add_implications}{\code{formal_context$add_implications()}}
\item \href{#method-compute_concepts}{\code{formal_context$compute_concepts()}}
\item \href{#method-extract_implications_concepts}{\code{formal_context$extract_implications_concepts()}}
\item \href{#method-convert_to_transactions}{\code{formal_context$convert_to_transactions()}}
\item \href{#method-export_implications_to_arules}{\code{formal_context$export_implications_to_arules()}}
\item \href{#method-save}{\code{formal_context$save()}}
\item \href{#method-load}{\code{formal_context$load()}}
\item \href{#method-dim}{\code{formal_context$dim()}}
\item \href{#method-print}{\code{formal_context$print()}}
\item \href{#method-plot_lattice}{\code{formal_context$plot_lattice()}}
\item \href{#method-plot_context}{\code{formal_context$plot_context()}}
\item \href{#method-get_concept_support}{\code{formal_context$get_concept_support()}}
\item \href{#method-get_implication_support}{\code{formal_context$get_implication_support()}}
\item \href{#method-clone}{\code{formal_context$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creator for the Fomal Context class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$new(I, remove_const = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{I}}{(numeric matrix) The table of the formal context.}

\item{\code{remove_const}}{(logical) If \code{TRUE}, remove constant columns. The default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Columns of \code{I} must be named, and are the names of the attributes of the formal context.

If no \code{I} is used, the resulting \code{FormalContext} will be empty and not usable unless for loading a previously saved one.
}

\subsection{Returns}{
An object of the \code{FormalContext} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_empty"></a>}}
\subsection{Method \code{is_empty()}}{
Check if FormalContext is empty
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$is_empty()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the \code{FormalContext} is empty, that is, has not been provided with a matrix, and \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_implications"></a>}}
\subsection{Method \code{add_implications()}}{
Add a precomputed implication set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$add_implications(impl_set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{impl_set}}{(\code{ImplicationSet} object) The implications to add to this formal context.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, just updates the internal \code{implications} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_concepts"></a>}}
\subsection{Method \code{compute_concepts()}}{
Use Ganter Algorithm to compute concepts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$compute_concepts(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{(logical) TRUE will provide a verbose output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with all the concepts in the formal context.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_implications_concepts"></a>}}
\subsection{Method \code{extract_implications_concepts()}}{
Use modified Ganter algorithm to compute both concepts and implications
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$extract_implications_concepts(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{(logical) TRUE will provide a verbose output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, just updates the internal fields \code{concepts} and \code{implications}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert_to_transactions"></a>}}
\subsection{Method \code{convert_to_transactions()}}{
Convert the formal context to object of class \code{transactions} from the \code{arules} package
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$convert_to_transactions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{transactions} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_implications_to_arules"></a>}}
\subsection{Method \code{export_implications_to_arules()}}{
Export implications to \code{arules} format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$export_implications_to_arules(quality = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quality}}{(logical) Compute the interest measures for each rule?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{rules} object from \code{arules} package. If \code{quality == TRUE}, the function \code{interestMeasure} from \code{arules} is used.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\subsection{Method \code{save()}}{
Save a \code{FormalContext} to RDS format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$save(filename = tempfile(fileext = ".rds"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{(character) Path of the RDS file where to store the \code{FormalContext}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly the current \code{FormalContext}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\subsection{Method \code{load()}}{
Load a \code{FormalContext} from a RDS file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formal_context$load(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{(character) Path of the RDS file to load the \code{FormalContext} from.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The loaded \code{FormalContext}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dim"></a>}}
\subsection{Method \code{dim()}}{
Dimensions of the formal context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$dim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector with (number of objects, number of attributes).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Prints the formal context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Prints information regarding the formal context.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_lattice"></a>}}
\subsection{Method \code{plot_lattice()}}{
Plot the concept lattice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$plot_lattice(minsupp = 0, object_names = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{minsupp}}{(numeric) Only plot the sublattice generated with those concepts whose support is greater than or equal to \code{minsupp}.}

\item{\code{object_names}}{(logical) If \code{TRUE}, plot object names, otherwise omit them from the diagram.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, just plots the graph of the concept lattice.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_context"></a>}}
\subsection{Method \code{plot_context()}}{
Plot the formal context table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$plot_context()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, just plots the formal context.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_concept_support"></a>}}
\subsection{Method \code{get_concept_support()}}{
Get support of each concept
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$get_concept_support()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector with the support of each concept.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_implication_support"></a>}}
\subsection{Method \code{get_implication_support()}}{
Compute support of each implication
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$get_implication_support()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector with the support of each implication
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FormalContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
