% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implication_set.R
\docType{data}
\name{implication_set}
\alias{implication_set}
\title{R6 Class for Set of implications}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
implication_set
}
\arguments{
\item{name}{(character string) Optional name of the implication set.}

\item{attributes}{(character vector) Vector of names of the attributes on which we define the implications.}

\item{lhs}{(a \code{dgCMatrix}) Optional initial LHS of the implications stored.}

\item{rhs}{(a \code{dgCMatrix}) Optional initial RHS of the implications stored.}

\item{arules_imp}{(\code{rules} object) The implications obtained with \code{arules} to be imported.}

\item{lhs}{(a sparse matrix) The LHS of the rule to add.}

\item{rhs}{(a sparse matrix) The RHS of the rule to add.}

\item{implications}{(\code{ImplicationSet}) The set of implications to append at the end of the current set.}

\item{S}{(a \code{SparseSet} object)  Fuzzy set to compute its closure. Use class \code{SparseSet} to build it.}

\item{reduce}{(logical) Reduce the implications using simplification logic?}

\item{verbose}{(logical) Show verbose output?}

\item{S}{(a vector) Vector with the grades of each attribute (a fuzzy set).}

\item{attribute_filter}{(character vector) Names of the attributes to get recommendation for.}

\item{rules}{(character vector) Names of the rules to use. See \code{details}.}

\item{batch_size}{(integer) If the number of rules is large, apply the rules by batches of this size.}

\item{parallelize}{(logical) If possible, should we parallelize the computation among different batches?}

\item{reorder}{(logical) Should the rules be randonly reordered previous to the computation?}

\item{ncols}{(integer) Number of columns for the output.}

\item{lhs}{(character vector) Names of the attributes to filter the LHS by. If \code{NULL}, no filtering is done on the LHS.}

\item{rhs}{(character vector) Names of the attributes to filter the RHS by. If \code{NULL}, no filtering is done on the RHS.}

\item{drop}{(logical) Remove the rest of attributes in RHS?}

\item{idx}{(integer vector) Indices of the implications to extract.}

\item{idx}{(integer vector) Indices of implications to remove.}
}
\value{
A new \code{implication_set} object.

A character vector with the names of the attributes used in the implications.

Nothing, just updates the internal matrices for LHS and RHS.

A \code{rules} object as used by package \code{arules}.

Nothing, just updates the internal matrices for LHS and RHS.

Nothing, just updates the internal matrices for LHS and RHS.

The cardinality of the implication set.

\code{TRUE} if the set of implications is empty, \code{FALSE} otherwise.

A vector with two components: the number of attributes present in each of the LHS and RHS of each implication in the set.

If \code{reduce == FALSE}, the output is a fuzzy set corresponding to the closure of \code{S}. If \code{reduce == TRUE}, a list with two components: \code{closure}, with the closure as above, and \code{implications}, the reduced set of implications.

A fuzzy set describing the values of the attributes in \code{attribute_filter} within the closure of \code{S}.

Nothing, just updates the internal matrices for LHS and RHS.

A string with all the implications in the set.

A string in LaTeX format that prints nicely all the implications.

A sparse matrix representing the LHS of the implications in the set.

A sparse matrix representing the RHS of the implications in the set.

An \code{ImplicationSet} that is a subset of the current set, only with those rules which has the attributes in \code{lhs} and \code{rhs} in their LHS and RHS, respectively.

A new \code{ImplicationSet} with only the rules given by the \code{idx} indices.

An \code{ImplicationSet} with the implications at given indices removed with respect to the original set.
}
\description{
This class implements the structure needed to store implications and the methods associated.

Initialize with an optional name

Get the names of the attributes

Import from arules object

Convert to arules format

Add new implication to the set

Append implications to the current ones.

Cardinality: Number of implications in the set

Empty set

Size: number of attributes in each of LHS and RHS

Compute the semantic closure of a fuzzy set with respect to the implication set

Generate a recommendation for a subset of the attributes

Apply rules to remove redundancies

Print all implications to text

Export to LaTeX

Get internal LHS matrix

Get internal RHS matrix

Filter implications by attributes in LHS and RHS

Get a subset of the implication set

Remove implications from the set
}
\details{
Currently, the implemented rules are \code{"generalization"}, \code{"simplification"}, \code{"reduction"} and \code{"composition"}.
}
\keyword{datasets}
