% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{select_repository_context}
\alias{select_repository_context}
\title{GUI to select and download a context from the repository}
\usage{
select_repository_context(meta)
}
\arguments{
\item{meta}{A list of metadata objects (obtained via \code{get_fcarepository_contexts}).}
}
\description{
GUI to select and download a context from the repository
}
