% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_concept.R
\name{sparse_concept}
\alias{sparse_concept}
\title{R6 class for a fuzzy concept with sparse internal representation}
\description{
This class implements the data structure and methods for fuzzy concepts.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{sparse_concept$new()}}
\item \href{#method-get_extent}{\code{sparse_concept$get_extent()}}
\item \href{#method-get_intent}{\code{sparse_concept$get_intent()}}
\item \href{#method-print}{\code{sparse_concept$print()}}
\item \href{#method-to_latex}{\code{sparse_concept$to_latex()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Creator for objects of class \code{SparseConcept}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_concept$new(extent, intent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{extent}}{(\code{SparseSet}) The extent of the concept.}

\item{\code{intent}}{(\code{SparseSet}) The intent of the concept.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{SparseConcept}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_extent"></a>}}
\subsection{Method \code{get_extent()}}{
Internal \code{SparseSet} for the extent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_concept$get_extent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The \code{SparseSet} representation of the extent.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_intent"></a>}}
\subsection{Method \code{get_intent()}}{
Internal \code{SparseSet} for the intent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_concept$get_intent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The \code{SparseSet} representation of the intent.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Prints the concept to console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_concept$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A string with the elements of the set and their grades between brackets {}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_latex"></a>}}
\subsection{Method \code{to_latex()}}{
Write the concept in LaTeX format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sparse_concept$to_latex()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The fuzzy concept in LaTeX.
}
}
}
