% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_set.R
\docType{data}
\name{sparse_set}
\alias{sparse_set}
\title{R6 class for a fuzzy set with sparse internal representation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sparse_set
}
\arguments{
\item{attributes}{(character vector) Names of the attributes that will be available in the fuzzy set.}

\item{M}{(numeric vector or column \code{Matrix}) Values (grades) to be assigned to the attributes.}

\item{attributes}{(character vector) Names of the attributes to assign a grade to.}

\item{values}{(numeric vector) Grades to be assigned to the previous \code{attributes}.}
}
\value{
An object of class \code{SparseSet}.

The internal sparse \code{Matrix} representation of the set.

A character vector with the names of the attributes.

The number of attributes that are defined for this fuzzy set.

A string with the elements of the set and their grades between brackets {}.

The fuzzy set in LaTeX.
}
\description{
This class implements the data structure and methods for fuzzy sets.

Creator for objects of class \code{SparseSet}

Assign grades to attributes in the set

Internal \code{Matrix}

Attributes defined for the set

Number of attributes

Prints the set to console

Write the set in LaTeX format
}
\details{
If \code{M} is omitted, the fuzzy set is the empty set. Later, one can use the \code{assign} method to assign grades to any of its attributes.
}
\keyword{datasets}
