% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptlist.R
\name{to_latex}
\alias{to_latex}
\alias{to_latex.conceptlist}
\title{Print a list of concepts to LaTeX}
\usage{
to_latex(x)

\method{to_latex}{conceptlist}(x)
}
\arguments{
\item{x}{A \code{conceptlist} as result of subsetting a \code{ConceptLattice}}
}
\value{
Nothing, just prints a list of concepts in LaTeX nicely.
}
\description{
Print a list of concepts to LaTeX
}
\examples{
# Build a formal context
fc_planets <- FormalContext$new(planets)

# Find its concepts
fc_planets$find_concepts()

# Print the first 3 concepts to latex
to_latex(fc_planets$concepts[1:3])

}
